+++ source/Core/Price.php
@@ -47,14 +47,12 @@ class Price
      */
     protected $_dVat = 0.0;
 
-
     /**
      * Assigned discount array
      *
      * @var array
      */
     protected $_aDiscounts = null;
-
 
     /**
      * Price entering mode
@@ -125,7 +123,7 @@ class Price
      */
     public function setVat($dVat)
     {
-        $this->_dVat = (double) $dVat;
+        $this->_dVat = (float) $dVat;
     }
 
     /**
@@ -141,9 +139,9 @@ class Price
     public function setUserVat($newVat)
     {
         if (!$this->isNettoMode() && $newVat != $this->_dVat) {
-            $this->_dBrutto = self::Netto2Brutto(self::Brutto2Netto($this->_dBrutto, $this->_dVat), (double) $newVat);
+            $this->_dBrutto = self::Netto2Brutto(self::Brutto2Netto($this->_dBrutto, $this->_dVat), (float) $newVat);
         }
-        $this->_dVat = (double) $newVat;
+        $this->_dVat = (float) $newVat;
     }
 
     /**
@@ -355,7 +353,7 @@ class Price
      */
     public static function percent($dValue, $dPercent)
     {
-        return ((double) $dValue * (double) $dPercent) / 100.0;
+        return ((float) $dValue * (float) $dPercent) / 100.0;
     }
 
     /**
@@ -378,7 +376,7 @@ class Price
             return 0;
         }
 
-        return (double) ((double) $dBrutto * 100.0) / (100.0 + (double) $dVat);
+        return (float) ((float) $dBrutto * 100.0) / (100.0 + (float) $dVat);
     }
 
     /**
@@ -393,7 +391,7 @@ class Price
      */
     public static function netto2Brutto($dNetto, $dVat)
     {
-        return (double) $dNetto + self::percent($dNetto, $dVat);
+        return (float) $dNetto + self::percent($dNetto, $dVat);
     }
 
     /**
@@ -407,9 +405,8 @@ class Price
     {
         $oCur = \OxidEsales\Eshop\Core\Registry::getConfig()->getActShopCurrencyObject();
 
-        return ((double) $dPrice) * $oCur->rate;
+        return ((float) $dPrice) * $oCur->rate;
     }
-
 
     /**
      * Sets discount to price
